Added LightGroup object class, same interface as Light but for controlling groups
The LightGroup class implements an identical interface as Light (in fact
it is a subclass derived from Light) but it works with the light group
API to control multiple lights at once.

There is also an AllLights class that enables control of all lights at
once. This makes use of the undocumented (even in the official API!)
feature that "group 0" is all lights connected to the bridge.