Streamline code around PY3K, isinstance, api_paths
This is the new version of edits redone against the current code base.  This Pull Request is a full replacement of PR #30

In `get_group()` there is a check to see if `group_id` is False followed by a check to see if `group_id` is None.  In Python, it is considered to be risky to depend on the difference between False and None.