add initial scene support
The following adds some basic scene support, including a Scene base
object, the ability to list all scenes. The ability to activate a
scence by group_id and scene_id, and a convenience method to activate
a scene based on scene name and group name.

It is intended to be the basis for making automation rules in home
assistant be easier to write.