make code pass flake8 (mostly)
This makes the code largely flake8 compliant, with some exceptions.

80 col line length is skipped. Line length conversation can happen
later.

The 'noqa' stanza is used to handle the use of python2 functions that
don't exist in python3. The typically handling for this is to use six,
however as a single file module with no dependencies, this isn't all
that simple.

Travis pep8 test running is enabled for this as well.