build framework for testing more complex API interactions
This builds a framework for having a set of canned test responses
based on HTTP Requests. It does this by monkey patching out
HTTPConnection, and redirecting it into a static multi level
dictionary. This allows complicated flows to be tested against known
sample data, and ensure that we're actually testing the various calls
back to the server.

A simple use of this in getting the data for one of the lights is
added. This will be expanded from here once it's accepted as an
approach.